fix name token auth failed
