+++ src/main/java/com/webank/wecross/account/AccountManager.java
@@ -4,7 +4,6 @@ import static com.webank.wecross.exception.WeCrossException.ErrorCode.GET_UA_FAI
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.webank.wecross.exception.WeCrossException;
-import com.webank.wecross.network.client.ClientConnection;
 import com.webank.wecross.network.client.ClientMessageEngine;
 import com.webank.wecross.network.client.Request;
 import com.webank.wecross.network.client.Response;
@@ -25,7 +24,6 @@ public class AccountManager {
     private ClientMessageEngine engine;
 
     private UniversalAccountFactory universalAccountFactory;
-    private ClientConnection clientConnection;
 
     private RouterLoginAccountContext adminContext;
     private AccountSyncManager accountSyncManager;
@@ -36,7 +34,6 @@ public class AccountManager {
     private Map<String, UniversalAccount> token2UA = new ConcurrentHashMap<>();
 
     public void start() {
-        logger.info("clientConnection: {}", clientConnection);
         timer.schedule(
                 new TimerTask() {
                     @Override
@@ -141,14 +138,6 @@ public class AccountManager {
 
     public void setAccountSyncManager(AccountSyncManager accountSyncManager) {
         this.accountSyncManager = accountSyncManager;
-    }
-
-    public ClientConnection getClientConnection() {
-        return clientConnection;
-    }
-
-    public void setClientConnection(ClientConnection clientConnection) {
-        this.clientConnection = clientConnection;
     }
 
     public void setTimer(Timer timer) {
