+++ src/main/java/com/webank/wecross/config/AuthFilterConfig.java
@@ -1,5 +1,6 @@
 package com.webank.wecross.config;
 
+import com.webank.wecross.network.client.ClientConnection;
 import com.webank.wecross.network.client.ClientMessageEngine;
 import com.webank.wecross.network.rpc.authentication.AuthFilter;
 import com.webank.wecross.network.rpc.authentication.RemoteAuthFilter;
@@ -12,10 +13,14 @@ public class AuthFilterConfig {
     @Resource(name = "newAccountManagerEngine")
     ClientMessageEngine accountManagerEngine;
 
+    @Resource(name = "newClientConnection")
+    ClientConnection clientConnection;
+
     @Bean
     public AuthFilter newAuthFilter() {
         RemoteAuthFilter remoteAuthFilter = new RemoteAuthFilter();
         remoteAuthFilter.setRemoteEngine(accountManagerEngine);
+        remoteAuthFilter.setClientConnection(clientConnection);
         remoteAuthFilter.registerAuthUri("/auth/register");
         remoteAuthFilter.registerAuthUri("/auth/login");
         remoteAuthFilter.registerAuthUri("/auth/logout");
